Update use-cases.md
Expand the use cases document to introduce and define the concept capabilities, and create a space for the enumeration of relevant capabilities.

Following the discussion regarding native attestation mechanisms as a source of truth (see https://github.com/antifraudcg/proposals/issues/8), I'd like to suggest clarifying the functional requirements for our key use cases. I'm suggesting expanding the use cases document to introduce the concept of "capabilities," allowing us to enumerate the types of inferences we need to make in order to detect the relevant types of fraud and abuse. Following this enumeration, we can hopefully parallelize the discussion of these capabilities, their unique complexities, and ultimately solicit designs to avail these capabilities subject to the agreed upon requirements.